requirements.txt
absl-py==2.1.0
acvl-utils==0.2
annotated-types==0.7.0
anyio==4.2.0
arel==0.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-asgi-testclient==1.4.11
async-lru==2.0.4
attrs==24.1.0
Automat==20.2.0
azure-core==1.30.2
azure-storage-blob==12.22.0
Babel==2.15.0
batchgenerators==0.25
batchgeneratorsv2==0.2.1
bcrypt==3.2.0
beautifulsoup4==4.12.3
bidict==0.23.1
bleach==6.1.0
blinker==1.4
cacheout==0.14.1
cachetools==4.2.4
certifi==2020.6.20
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.0.3
cloud-init==24.1.3
colorama==0.4.4
comm==0.2.2
command-not-found==0.3
configobj==5.0.6
connected-components-3d==3.18.0
constantly==15.1.0
contourpy==1.2.1
cryptography==3.4.8
cycler==0.12.1
dbus-python==1.2.18
debugpy==1.8.3
decorator==5.1.1
defusedxml==0.7.1
dicom2nifti==2.5.0
distinctipy==1.3.4
distlib==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
dynamic-network-architectures==0.3.1
einops==0.8.0
exceptiongroup==1.2.2
executing==2.0.1
fastapi==0.109.0
fastjsonschema==2.20.0
ffmpeg-python==0.2.0
fft-conv-pytorch==1.2.0
filelock==3.15.4
fonttools==4.53.1
fqdn==1.5.1
fsspec==2024.6.1
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
giturlparse==0.12.0
grad-cam==1.5.3
graphviz==0.20.3
greenlet==3.1.0
grpcio==1.66.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.20.2
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.5
hyperlink==21.0.0
idna==3.3
imagecodecs==2024.9.22
imageio==2.34.2
importlib-metadata==4.6.4
imutils==0.5.4
incremental==21.3.0
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.3
isodate==0.6.1
isoduration==20.11.0
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.0.3
joblib==1.4.2
json5==0.9.25
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.11
keyring==23.5.0
kiwisolver==1.4.5
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.4
lightning-utilities==0.11.6
linecache2==1.0.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
more-itertools==8.10.0
mpmath==1.3.0
multidict==6.1.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==3.3
nibabel==5.2.1
nnunetv2==2.5.1
notebook==7.2.1
notebook_shim==0.2.4
numerize==0.12
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
opencv-python==4.10.0.84
overrides==7.7.0
packaging==24.1
pandas==2.1.4
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.8.0
pillow==10.2.0
platformdirs==4.2.2
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==3.20.3
psutil==5.9.8
psycopg-binary==3.2.1
psycopg2-binary==2.9.9
PTable==0.9.2
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydicom==2.4.4
pydub==0.25.1
Pygments==2.18.0
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pynrrd==0.4.3
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyparted==3.11.7
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.0
python-gdcm==3.0.24.1
python-json-logger==2.0.7
python-magic==0.4.27
python-multipart==0.0.5
pytz==2022.1
PyYAML==5.4.1
pyzmq==26.1.0
qtconsole==5.5.2
QtPy==2.4.1
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.8.1
rpds-py==0.19.1
safetensors==0.4.4
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.0
screen-resolution-extra==0.0.0
seaborn==0.13.2
SecretStorage==3.3.1
Send2Trash==1.8.3
service-identity==18.1.0
shapely==2.0.2
SimpleITK==2.4.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sos==4.5.6
soupsieve==2.5
SQLAlchemy==2.0.34
ssh-import-id==5.11
stack-data==0.6.3
starlette==0.35.1
stringcase==1.2.0
supervisely==6.73.187
sympy==1.13.1
systemd-python==234
tensorboard==2.17.1
tensorboard-data-server==0.7.2
terminado==0.18.1
threadpoolctl==3.5.0
tifffile==2024.7.24
timm==1.0.8
tinycss2==1.3.0
tokenizers==0.19.1
tomli==2.0.1
torch==2.4.0
torchmetrics==1.4.1
torchsummary==1.5.1
torchvision==0.19.0
torchxrayvision==1.2.3
tornado==6.4.1
tqdm==4.66.5
traceback2==1.4.0
traitlets==5.14.3
transformers==4.43.4
trimesh==3.23.5
triton==3.0.0
ttach==0.0.3
Twisted==22.1.0
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
ultralytics==8.2.92
ultralytics-thop==2.0.6
unattended-upgrades==0.1
unittest2==1.1.0
uri-template==1.3.0
urllib3==2.2.2
uvicorn==0.30.6
uvloop==0.20.0
varname==0.13.3
virtualenv==20.26.4
wadllib==1.3.6
WALinuxAgent==2.2.46
watchfiles==0.24.0
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==10.4
Werkzeug==3.0.4
widgetsnbextension==4.0.11
xkit==0.0.0
yacs==0.1.8
zipp==1.0.0
zope.interface==5.4.0
zstd==1.5.5.1
